ai-contador/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
